Produccion

+++ Django/don_galleto/insumos_app/models.py
@@ -6,6 +6,6 @@ class Insumo(models.Model):
     nombre = models.TextField(max_length=100)
     descripcion = models.TextField(max_length=500)
     medida = models.ForeignKey(Medida, on_delete=models.CASCADE, null=False, blank=False)
-    cantidad_minima = models.IntegerField(null=False, default=0)
+    cantidad_minima = models.DecimalField(decimal_places=3, max_digits=10, null=False, default=0)
     def __str__(self):
         return f"{self.nombre} / {self.medida.nombre}"
